refactor: Overhaul directory structure to be domain-driven
