+++ frontend/src/features/wish/components/WishEmptyList.tsx
