+++ frontend/src/features/wish/components/WishItem.tsx
