+++ frontend/src/features/wish/components/WishSearch.tsx
