+++ frontend/src/features/wish/components/index.ts
