+++ frontend/src/features/wish/types/Wish.ts
