+++ frontend/src/features/wish/types/WishHeaderProps.ts
