+++ frontend/src/features/wish/types/WishItemProps.ts
