+++ frontend/src/features/wish/types/WishSearchProps.ts
